+++ cvat-ui/src/components/projects-page/projects-page.tsx
@@ -11,6 +11,7 @@ import Spin from 'antd/lib/spin';
 import { CombinedState, ProjectsQuery } from 'reducers/interfaces';
 import { getProjectsAsync } from 'actions/projects-actions';
 import FeedbackComponent from 'components/feedback/feedback';
+import ImportDatasetModal from 'components/import-dataset-modal/import-dataset-modal';
 import EmptyListComponent from './empty-list';
 import TopBarComponent from './top-bar';
 import ProjectListComponent from './project-list';
@@ -22,16 +23,18 @@ export default function ProjectsPageComponent(): JSX.Element {
     const projectFetching = useSelector((state: CombinedState) => state.projects.fetching);
     const projectsCount = useSelector((state: CombinedState) => state.projects.current.length);
     const gettingQuery = useSelector((state: CombinedState) => state.projects.gettingQuery);
+    const isImporting = useSelector((state: CombinedState) => state.projects.restoring);
 
     const anySearchQuery = !!Array.from(new URLSearchParams(search).keys()).filter((value) => value !== 'page').length;
 
-    useEffect(() => {
+    const getSearchParams = (): Partial<ProjectsQuery> => {
         const searchParams: Partial<ProjectsQuery> = {};
         for (const [param, value] of new URLSearchParams(search)) {
             searchParams[param] = ['page', 'id'].includes(param) ? Number.parseInt(value, 10) : value;
         }
-        dispatch(getProjectsAsync(searchParams));
-    }, []);
+
+        return searchParams;
+    };
 
     useEffect(() => {
         const searchParams = new URLSearchParams();
@@ -46,6 +49,12 @@ export default function ProjectsPageComponent(): JSX.Element {
         });
     }, [gettingQuery]);
 
+    useEffect(() => {
+        if (isImporting === false) {
+            dispatch(getProjectsAsync(getSearchParams()));
+        }
+    }, [isImporting]);
+
     if (projectFetching) {
         return <Spin size='large' className='cvat-spinner' />;
     }
@@ -55,6 +64,7 @@ export default function ProjectsPageComponent(): JSX.Element {
             <TopBarComponent />
             {projectsCount ? <ProjectListComponent /> : <EmptyListComponent notFound={anySearchQuery} />}
             <FeedbackComponent />
+            <ImportDatasetModal />
         </div>
     );
 }